updated for version 7.3.797 Problem:    Compiler warning for size_t to int conversion. (Skeept) Solution:   Add type casts.

+++ src/ex_cmds.c
@@ -4764,8 +4764,8 @@ do_sub(eap)
 					 * substitute may have inserted or
 					 * deleted characters before the
 					 * cursor. */
-					len_change = STRLEN(new_line)
-							  - STRLEN(orig_line);
+					len_change = (int)STRLEN(new_line)
+						     - (int)STRLEN(orig_line);
 					curwin->w_cursor.col += len_change;
 					ml_replace(lnum, new_line, FALSE);
 				    }
